requirements.txt
PyQt6>=6.4.0
openai>=1.0.0
SpeechRecognition
pyttsx3
pyaudio>=0.2.13
elevenlabs
pydub==0.25.1
flask
anthropic>=0.3.0
openai-whisper>=20231117
deepgram-sdk>=2.11.0
sounddevice>=0.4.6
soundfile==0.12.1
numpy>=1.24.0
pyautogui==0.9.54
pyperclip>=1.8.2
PyYAML>=6.0.1
